connstat suna column doesn't seem to represent send+unacknowledged data #17 connstat should gracefully handle keyboard interrupt #15 connstat command can't handle running on an older kernel #18 would like to display TCP receive queue size with connstat #20
The main bug being fixed (#17) is caused by the caller of
record_ipv4_route() passing in the last two arguments in the wrong
order.

In addition to this simple fix, in order to prevent a similar error in
the future, record_ipv4_route() (which is renamed to record_ipv4_conn())
now takes a pointer to a pameter structure instead of a large number of
arguments.

This change also applies LLVM code style to the kernel module source,
and adds a GitHub Action to check python lint using flake8.